Combined controllers but it's not working for me. Added rest template builder in the topicserviceapplication.java, deleted bootstrap.yml, changed @repoistory annotation to @restrepositoryresource
