+++ hydra-topic-service/src/main/java/com/revature/HydraTopicServiceApplication.java
@@ -2,8 +2,13 @@ package com.revature;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.client.RestTemplate;
 
 @EnableCircuitBreaker
 @EnableEurekaClient
@@ -13,4 +18,11 @@ public class HydraTopicServiceApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(HydraTopicServiceApplication.class, args);
 	}
+	@LoadBalanced
+	@Bean
+	public RestTemplate buildRestTemplate(RestTemplateBuilder restTemplateBuilder) {
+		return restTemplateBuilder.build();
+	}
 }
+
+
