+++ hydra-topic-service/src/main/java/com/revature/repository/SubtopicNameRepository.java
@@ -1,15 +1,14 @@
 package com.revature.repository;
 
-import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
+import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+
 
 import com.revature.model.SubtopicName;
 
-@Repository
+@RepositoryRestResource
 public interface SubtopicNameRepository extends JpaRepository<SubtopicName, Integer> {
 	public SubtopicName findByid(Integer id);
 	public SubtopicName findByName(String name);
-	public List<SubtopicName> findAll();
 }
