+++ hydra-topic-service/src/main/java/com/revature/services/SubTopicService.java
@@ -118,7 +118,9 @@ public class SubTopicService {
 	  }
 
 	  public List<Subtopic> getSubtopics() {
+		 
 	    return subtopicRepository.findAll();
+	    
 	  }
 
 	  /**
