+++ hydra-topic-service/src/main/resources/application.properties
@@ -1,15 +1,12 @@
-server.port=8090
+server.port=8889
+spring.application.name=topic-service
+eureka.client.serviceUrl.defaultZone=${EUREKA_URI:http://localhost:8761/eureka}
+eureka.instance.preferIpAddress=true
 spring.datasource.url=jdbc:oracle:thin:@bamprod.cvoui7q38caj.us-east-2.rds.amazonaws.com:1521:ORCL
 spring.datasource.username=bamproduser
 spring.datasource.password=revabam2017
-#spring.datasource.url=jdbc:oracle:thin:@usfjavadb.cxzj3wk2fr9x.us-east-2.rds.amazonaws.com:1521:orcl
-#spring.datasource.username=project3
-#spring.datasource.password=p4ssw0rd
 spring.datasource.driver-class-name=oracle.jdbc.driver.OracleDriver
 spring.jpa.database-platform=org.hibernate.dialect.Oracle10gDialect
 spring.jpa.show-sql=false
 spring.jpa.hibernate.ddl-auto=validate
-
-eureka.client.fetch-registry=true
-eureka.client.register-with-eureka=true
