--- hydra-topic-service/src/main/resources/bootstrap.yml
@@ -1,6 +1,0 @@
-spring:
-  application:
-    name: topic-service
-  cloud:
-    config:
-      uri: http://localhost:8090/
